Initial Running Push
I tried to make a few of the .dart files more readable, but the sheer amount of nested objects makes it the equivalent of shoveling hay, at least on the onset. I'll try to give it a more proper go whenever I edit parts of the signup page/etc, unless there's a better way of organizing these files/viewing them I'm unware of.